IMPALA-9267: Fix DCHECK in ClientRequestState::UpdateNonErrorExecState
Fixes a DCHECK in ClientRequestState::UpdateNonErrorExecState where the
ClientRequestState ExecState attempts to transition from the ERROR to
the FINISHED state. The DCHECK was added in IMPALA-6894 in order to
prevent any invalid state transition attempts.

The fix is to modify UpdateNonErrorExecState so that it skips any
attempt to transition from the ERROR to the FINISHED state, which is in
line with the behavior prior to IMPALA-6894.

Testing:
* Ran core tests, unable to reproduce the original issue locally

Change-Id: Ie47444ed67704d9469310727eeec2e9a66516e77
Reviewed-on: http://gerrit.cloudera.org:8080/14991
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit a36e2c9c5eb313f5fddb08a52382ca3237397706)

Change-Id: Ic5165fb1986789ea5da5c9602584168799203978
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/66815
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>