IMPALA-9287: Add support for embedded metastore mode
This is an update to the earlier patch which was reverted for
IMPALA-9287. The original patch added dependency to hive-metastore
module in hive-3 profile which was unnecessary since we already have
a dependency to standalone-hive-metastore module.

Adding hive-metastore is not recommended since it brings along with
it hive-serde as well which has a conflicting version of flatbuffers.
This causes table loading errors on the catalog side.

Testing:
  1. Ran test_kudu_table_create_without_hms.py
  2. [WIP] Running core tests with CDP build.

Change-Id: Iab7de6a7f11ef0bff85c0fb03f004756c3cde784