Implements a basic active-passive leader election based off a file
In order to test the logic, run the test FileSystemUtil#testLeaderElection
using mvn. The test never completes since it runs two never-ending leaderElector
threads. In order to see what is happening look into the file in
logs/fe_tests/org.apache.impala.common.FileSystemUtilTest-output.txt.

The main idea is the leader updates the file at a faster rate than non-leaders.
When the leader dies, non-leaders eventually realize that file is not being updated
and hence take over. The code doesn't use local-timestamps and operates on the
concept of duration. Needs more investigation on tolerant is this with respect
to clock skews.

the code injects random leader failures and recoveries to test fail-over.

The code probably needs a thorough review to see if there are race conditions.

I think we can replace the rename to a createFile with overwrite flag set to false
provide its atomic operation accross.

Change-Id: Ic6a7d708bcf5f7ac81c045a7147afe8d57334f3a