IMPALA-9824: MetastoreClientPool should be singleton
This patch is a refactor-only patch. It changes the MetastoreClientPool
into process wide singleton object. Currently, it is possible to
instantiate multiple MetastoreClientPools although it doesn't seem to
be a problem since all the places which instantiate it happen to run
in separate processes. It would be better to enforce this in code.
Making MetastoreClientPool a singleton also makes it easier to access
from anywhere in the code without the need to explicitly pass its
reference around which complicates the code flow unnecessarily.

Additionally, this patch also introduces two new configurations
coordinator_initial_hms_connections and catalog_initial_hms_connections
which define the initial number of HMS clients which are created
for the respective services. The default values of both the
configurations are kept same as previous values from the code.

Testing:
1. No new tests were added since this was mostly a refactor-only patch.
2. Ran existing core tests.

Change-Id: I5477cba68b9053dfae8e7ed785f4d7519c12cd0f