IMPALA-10502: Handle CREATE/DROP events correctly
The current way to detect self-events in case of CREATE/DROP events on
database, table and partition is problematic when the same object is
created and dropped repeatedly in quick succession. This happens mainly
due to couple of reasons. For example if we have a sequence of
CREATE_TABLE, DROP_TABLE, CREATE_TABLE ... on the same table, it is
possible that when the create table event is being processed, the table
is not present in catalog because it was dropped recently. In such a
case, events processor does not have enough state information in
catalogd to determine that this table has been dropped from the
catalogd and the event should be ignored. Similarly, if a drop event
is being processed, it is possible that the table has been recreated
with the same name when the drop event is received. In such a case,
events processor removes the table from the catalogd.

This can cause problems for queries which expect the table to exist or
not exist. E.g create table query fails with a table already exists or
a drop table query fails with table does not exist error.

In order to fix this issue, catalogd now keeps track of dropped objects
in a deleteLog which are garbage collected as the events come in. Every
time a database, table or parition is dropped, the deleteLog is
populated with the the drop event id generated due to the drop
operation. This deleteLog is looked up when the event is received to
determine if the event can be ignored.

Testing:
1. Added a new test which loops to create create/drop events for
database, table and partitions.
2. Added new metrics which the test verifies to ensure that events
don't create or drop the object.

Change-Id: Ia2c5e96b48abac015240f20295b3ec3b1d71f24a