IMPALA-10502: Fetch events in batches (Addendum)
The earlier change for IMPALA-10502 passes in a batch size
of -1 to fetch all the events from a given event id. While this works when
HMS backing database is postgres, it doesn't work well when the HMS backend
is a MySQL database due to HIVE-20226. This change workaround the hive
bug to fetch the events in batches of 1000 instead of fetching all the events
in one RPC during the DDL execution.

Testing:
1. Added a unit test for the new changes introduced.
2. Ran the previously failing tests on MySQL HMS backend.

Change-Id: Ia8dd2a4f0644e8bb713e43191fb69a56f6fd0ad6